Move all room related servlets into rest/admin/rooms.py
